+++ src/database/postgres/model_food.rs
@@ -75,7 +75,7 @@ ORDER BY
 }
 
 /// Used to skip serializtion if value is None or false
-#[expect(clippy::trivially_copy_pass_by_ref)]
+#[expect(clippy::trivially_copy_pass_by_ref, clippy::ref_option)]
 fn none_or_false(x: &Option<bool>) -> bool {
     if let Some(value) = x {
         return !value;