+++ src/main.rs
@@ -86,5 +86,8 @@ async fn main() -> Result<(), ApiError> {
     let postgres = database::db_postgres::db_pool(&app_env).await?;
     let redis = database::DbRedis::get_pool(&app_env).await?;
     BackupSchedule::init(&app_env);
-    api::serve(app_env, postgres, redis).await
+    tokio::spawn(api::serve(app_env, postgres, redis))
+        .await
+        .ok();
+    Ok(())
 }